Reorganizing functions needed for gradient checks
