+++ jcm/physics/speedy/large_scale_condensation_test.py
@@ -3,6 +3,7 @@ import jax.numpy as jnp
 import numpy as np
 import jax
 import functools
+from jax.test_util import check_vjp, check_jvp
 
 class TestLargeScaleCondensationUnit(unittest.TestCase):
 
@@ -88,8 +89,7 @@ class TestLargeScaleCondensationUnit(unittest.TestCase):
 
     # Gradient test
     def test_get_large_scale_condensation_tendencies_realistic_gradient_check(self):
-        from jax.test_util import check_vjp, check_jvp
-        from jax.tree_util import tree_map
+        from jcm.utils import convert_back, convert_to_float
         xy = (ix,il)
         zxy = (kx,ix,il)
         psa = jnp.ones((ix, il)) * 1.0110737
@@ -103,28 +103,6 @@ class TestLargeScaleCondensationUnit(unittest.TestCase):
         state = PhysicsState.zeros(zxy, specific_humidity=qa[:, jnp.newaxis, jnp.newaxis],normalized_surface_pressure=psa)
         physics_data = PhysicsData.zeros(xy, kx, humidity=humidity, convection=convection)
         boundaries = BoundaryData.ones(xy)
-
-        # Converting functions
-        def check_type_convert_to_float(x): # Do error catch block
-            try:
-                return x.astype(jnp.float32)
-            except AttributeError:
-                return jnp.float32(x)
-        def convert_to_float(x): 
-            return tree_map(check_type_convert_to_float, x)
-        def check_type_convert_back(x, x0):
-            try: 
-                if x0.dtype == jnp.float32:
-                    return x
-                else:
-                    return x0
-            except AttributeError:
-                if type(x0) == jnp.float32:
-                    return x
-                else:
-                    return x0
-        def convert_back(x, x0):
-            return tree_map(check_type_convert_back, x, x0)
 
         # Set float inputs
         physics_data_floats = convert_to_float(physics_data)
