+++ jcm/utils.py
@@ -1,5 +1,6 @@
 import jax.numpy as jnp
 from jax import jit
+from jax.tree_util import tree_map
 from dinosaur.coordinate_systems import HorizontalGridTypes
 
 # Function to take a field in grid space and truncate it to a given wavenumber
@@ -29,3 +30,27 @@ def pass_fn(operand):
 def ones_like(x):
     import jax.tree_util as jtu
     return jtu.tree_map(lambda y: jnp.ones_like(y), x)
+
+# Convert object to float 
+def check_type_convert_to_float(x): 
+    try:
+        return x.astype(jnp.float32)
+    except AttributeError:
+        return jnp.float32(x)
+def convert_to_float(x): 
+    return tree_map(check_type_convert_to_float, x)
+
+# Revery object with type float back to true type
+def check_type_convert_back(x, x0):
+    try: 
+        if x0.dtype == jnp.float32:
+            return x
+        else:
+            return x0
+    except AttributeError:
+        if type(x0) == jnp.float32:
+            return x
+        else:
+            return x0
+def convert_back(x, x0):
+    return tree_map(check_type_convert_back, x, x0)
